Initial Project
I started this project by creating 3 different apps:

InventoryItems:
   To house all data about the medical devices.

Profile:
   A proxy for django.contrib.auth.user, contains information about any employee. I created a django signal which generates a profile any time a User object is added.

ServiceCenter:
   Contains information about each individual service center that the client operates. This is intended for use in a business where there are multiple physical locations, and equipment can be shared between service centers. Each service center has a foreignkey relationship with Profile to determine which Manager should do the approval.

Set Groups: Employee, Manager, Admin, with appropriate permissions for each group.

Models:

Profile: A basic user profile for employees

InventoryEquipmentType: Used to describe the type of medical devices that the client supports

DeviceInstance: The representation of a single medical device, which has a unique Serial Number, and needs to be tracked for regulatory reasons. Protected data that can only be edited by a Service Center Manager.

Transaction: Employees cannot directly edit a DeviceInstance. They can only submit Transactions, which, when approved by a manager, will update the DeviceInstance information.
If a transaction is declined by manager, it will generate a new, identical transaction, and send it back to the employee that submitted to be corrected.