TransactionForm added, urls mapped
Created a modelform for employees to submit a transaction for review. The transaction_id and transaction_date are generated server-side, and not input into the form by the employee.

Mapped urls for the InventoryItems app. The endpoints are: /list/, /details/, /transactions/, and /submit/

most of these are generic list/detailviews, however the submit page displays the TransactionForm to be filled out by an employee.

Created basic templates to visualize the app layout, edited the settings.py to add a template folder and a static assets folder.