Added transaction approval
Transactions can now be submitted through the /submit/ endpoint.

Once these are submitted, the "approver" can monitor their pending transactions at the /approve/ endpoint. This page displays a list of all transactions that have been sent.

Once approved, the transaction will be set to approved=True, and it will no longer be displayed to the manager at the /approve/ endpoint. This will also update the corresponding DeviceInstance model, and save the model with the updated information from the transaction.

TODO:
1.Authentication/Login/Logout

2. /pending/ endpoint for users who have had their Transaction declined by the manager.